move equifax package to common service/rabbitmq
