+++ ratnaafin/src/main/java/com/ratnaafin/crm/common/rabbitmq/Equifax/CustomException/ConfigurationErrorException.java
@@ -1,4 +1,4 @@
-package com.ratnaafin.crm.common.service.rabbitMQ.Equifax.CustomException;
+package com.ratnaafin.crm.common.rabbitmq.Equifax.CustomException;
 
 public class ConfigurationErrorException extends Exception {
     public ConfigurationErrorException(String fromEntity,int configCode,String desc){
