+++ ratnaafin/src/main/java/com/ratnaafin/crm/common/rabbitmq/Equifax/CustomException/ServiceErrorException.java
@@ -1,4 +1,4 @@
-package com.ratnaafin.crm.common.service.rabbitMQ.Equifax.CustomException;
+package com.ratnaafin.crm.common.rabbitmq.Equifax.CustomException;
 
 public class ServiceErrorException extends  Exception{
     public ServiceErrorException(String errorMessage){
